+++ test/Foil.chocoV2.tests.ps1
@@ -33,7 +33,7 @@ Describe "Chocolatey V2 DSC-compliant package installation and uninstallation" {
 		It 'silently installs the latest version of a package' {
 			Install-ChocoPackage -Name $package -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds the locally installed package just installed' {
+		It 'detects the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 		It 'silently uninstalls the locally installed package just installed' {
@@ -54,7 +54,7 @@ Describe "Chocolatey V2 DSC-compliant package installation and uninstallation" {
 		It 'correctly passed parameters to the package' {
 			Get-ChildItem -Path (Join-Path -Path $env:ProgramFiles -ChildPath $package) -ErrorAction SilentlyContinue | Should -Not -BeNullOrEmpty
 		}
-		It 'finds the locally installed package just installed' {
+		It 'detects the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 		It 'silently uninstalls the locally installed package just installed' {
@@ -72,7 +72,7 @@ Describe "Chocolatey V2 DSC-compliant package installation and uninstallation" {
 		It 'silently installs the latest version of a package' {
 			Install-ChocoPackage -Name $package -PreRelease -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds the locally installed package just installed' {
+		It 'detects the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 		It 'silently uninstalls the locally installed package just installed' {
@@ -90,7 +90,7 @@ Describe "Chocolatey V2 pipline-based package installation and uninstallation" {
 		It 'searches for and silently installs the latest version of a package' {
 			Find-ChocoPackage -Name $package | Install-ChocoPackage -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds and silently uninstalls the locally installed package just installed' {
+		It 'detects and silently uninstalls the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 	}
@@ -102,7 +102,7 @@ Describe "Chocolatey V2 pipline-based package installation and uninstallation" {
 		It 'searches for and silently installs the latest version of a package' {
 			Find-ChocoPackage -Name $package | Install-ChocoPackage -Force | Should -HaveCount 3
 		}
-		It 'finds and silently uninstalls the locally installed package just installed, along with its dependencies' {
+		It 'detects and silently uninstalls the locally installed package just installed, along with its dependencies' {
 			Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage -RemoveDependencies | Should -HaveCount 3
 		}
 	}
@@ -114,7 +114,7 @@ Describe "Chocolatey V2 pipline-based package installation and uninstallation" {
 		It 'searches for and silently installs the latest version of a package' {
 			Find-ChocoPackage -Name $package | Install-ChocoPackage -Force -Parameters '' | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds and silently uninstalls the locally installed package just installed' {
+		It 'detects and silently uninstalls the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 	}
@@ -129,7 +129,7 @@ Describe "Chocolatey V2 pipline-based package installation and uninstallation" {
 		It 'correctly passed parameters to the package' {
 			Get-ChildItem -Path (Join-Path -Path $env:ProgramFiles -ChildPath $package) -ErrorAction SilentlyContinue | Should -Not -BeNullOrEmpty
 		}
-		It 'finds and silently uninstalls the locally installed package just installed' {
+		It 'detects and silently uninstalls the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 	}
@@ -141,7 +141,7 @@ Describe "Chocolatey V2 pipline-based package installation and uninstallation" {
 		It 'searches for and silently installs the latest version of a package' {
 			Find-ChocoPackage -Name $package -PreRelease | Install-ChocoPackage -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds and silently uninstalls the locally installed package just installed' {
+		It 'detects and silently uninstalls the locally installed package just installed' {
 			Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 		}
 	}
@@ -165,9 +165,9 @@ Describe "Chocolatey V2 multi-source support" {
 		{ Register-ChocoSource -Name $altSource -Location $altLocation | Where-Object {$_.Name -eq $altSource} } | Should -Not -Throw
 	}
 	It 'searches for and installs the latest version of a package from an alternate source' {
-		Find-ChocoPackage -Name $package -source $altSource | Install-ChocoPackage -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
+		Find-ChocoPackage -Name $package -Source $altSource | Install-ChocoPackage -Force | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 	}
-	It 'finds and uninstalls a package installed from an alternate source' {
+	It 'detects and uninstalls a package installed from an alternate source' {
 		Get-ChocoPackage -Name $package -LocalOnly -Exact | Uninstall-ChocoPackage | Where-Object {$_.Name -contains $package} | Should -Not -BeNullOrEmpty
 	}
 	It 'unregisters an alternative package source' {
@@ -190,7 +190,7 @@ Describe "Chocolatey V2 version filters" {
 		It 'searches for and silently installs a specific package version' {
 			Find-ChocoPackage -Name $package -Version $version -Exact | Install-ChocoPackage -Force | Where-Object {$_.Name -contains $package -and $_.Version -eq $version} | Should -Not -BeNullOrEmpty
 		}
-		It 'finds and silently uninstalls a specific package version' {
+		It 'detects and silently uninstalls a specific package version' {
 			Get-ChocoPackage -Name $package -Version $version -LocalOnly | UnInstall-ChocoPackage -Force | Where-Object {$_.Name -contains $package -and $_.Version -eq $version} | Should -Not -BeNullOrEmpty
 		}
 	}